Move value table folder
