+++ Source/Client/Core/UnitTests/gen/ValueTable/ValueTableReaderTests.gen.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "Contracts/ValueTableReaderTests.h"
+#include "ValueTable/ValueTableReaderTests.h"
 
 TestState RunValueTableReaderTests() 
 {
